* 修复上一个问题，缺少@service ， 这么坑。。

+++ src/main/java/com/jnote/mapper/JnoteBlogMapper.java
@@ -18,6 +18,6 @@ public interface JnoteBlogMapper {
 
     int updateByPrimaryKey(JnoteBlog record);
     
-    JnoteBlog getBlogById(Long id);
+    JnoteBlog getBlogById(Integer id);
     
 }
